Conform to POSIX for all-year DST in TZ strings
zic now generates a POSIX-conforming TZ string for timezones
where all-year DST is predicted for the indefinite future.
This implements a suggestion by Michael Deckers in:
https://mm.icann.org/pipermail/tz/2020-February/028834.html
This change does not affect any existing tzdb zones,
as we prefer adjusting the standard time to all-year DST.
* zic.c (rule_cmp): Compare two rules to be the same
if they both extend into the indefinite future.
(stringzone): For DST all year, generate something like
"XXX3EDT4,0/0,J365/23" (which conforms to current POSIX) instead
of "EST5EDT4,0/0,J365/25" (which does not).  Eventually POSIX
should change to allow either form, but we might as well be
compatible with current POSIX when we can.