Improve fix for localtime bug with slim leaps
* localtime.c (tzparse): Replace arg PARSELB with arg BASEP,
and do the parselb calculations ourselves rather than
relying on the caller.  All callers changed.  Use a different
bound for timezone transitions than for leap seconds, since
we need the former to decide how to merge a TZif file's TZ
string transitions, and we also need the latter to know when
to stop extending the transition table.  This fixes a bug
when the last leap second is after the last explicit
timezone transition.